datastruct/vector: Add vector_set, and various other changes.
* Hoist element base calculation to VECTOR_INDEX in impl.h.
* Add vector_set. This copies the given value to the specified index.
- The value is assumed to be the width of the vector's type.
- The vector's 'used' value is set to the index, if the index is greater
  than the current used value.
* Make vector_length return a size_t.
* Make vector_width return a size_t (should have been this way initially).
* Make vector_get take a const vector_t *.
* Various documentation improvements.

+++ libraries/datastruct/vector/length.c
@@ -10,7 +10,7 @@
 
 #include "impl.h"
 
-int vector_length(const vector_t *v)
+size_t vector_length(const vector_t *v)
 {
   return v->used;
 }
