+++ libraries/datastruct/vector/width.c
@@ -10,7 +10,7 @@
 
 #include "impl.h"
 
-int vector_width(const vector_t *v)
+size_t vector_width(const vector_t *v)
 {
   return v->width;
 }
